fix missing gc obj info

+++ rpython/rlib/rvmprof/src/shared/vmprof_unix.h
@@ -85,7 +85,7 @@ int vmprof_register_virtual_function(char *code_name, intptr_t code_uid,
 RPY_EXTERN 
 int vmprof_sample_stack_now_gc_triggered(void);
 RPY_EXTERN 
-int vmprof_report_minor_gc_objs(double time_start, int * array_ptr, int array_size);
+int vmprof_report_minor_gc_objs(double time_start, intptr_t * array_ptr, int array_size);
 
 
 void vmprof_aquire_lock(void);
